Fixed hiding/showing of help floating button
Earlier events from the DownloadManager service were used to evaluate
whether Help Floating button should be shown or not. So, if service is
already running, it won't send those events to the activity which was
leads to wrong showing of help floating button. Now fragment itself is
connected to service and checks the download count.